Introduce chunking in the logs.
In the background logger we introduce chunking, a process when
we read a number of messages from the queue and process that chunk
and flush afterwards. It approach can lead to better performance,
for example if you have hLogStdout, and hFlush as a flush you may
may still use BufferredHandle instead of line one, without introducing
delays in writing logs.

Currently there are 2 ways to write consume log messages:
  1. old one (Just 1) in limit parameter of the Capacity - each message
     if written and processed one by one
  2. `Nothing` unbounded (actually bounded by the TQueue size)
  3. `Just i` bounded

Default choice is `Just 32`, it's an arbitrary decision that works in
practice for our company, but we have not spend much time investigating.